feature : refactored PageFallback. Added loaders folder to modules.
