+++ src/modules/loaders/ui/components/CircularLoaderContainer.jsx
@@ -1,0 +1,29 @@
+import Paper from '@mui/material/Paper';
+import CircularProgress from '@mui/material/CircularProgress';
+
+const CircularLoaderContainer = ({
+  containerVariant = 'outlined',
+  containerSX = {},
+  progressColor = 'primary',
+  progressSize = 60,
+}) => {
+  return (
+    <Paper
+      variant={containerVariant}
+      sx={{
+        display: 'flex',
+        flexDirection: 'column',
+        justifyContent: 'center',
+        alignItems: 'center',
+        height: '100%',
+        ...containerSX,
+      }}
+      role="status"
+      aria-label="Loading page"
+    >
+      <CircularProgress size={progressSize} color={progressColor} />
+    </Paper>
+  );
+};
+
+export default CircularLoaderContainer;
